Added draft to User control window
