+++ WpfClient/View/FilmEditWindow.xaml
@@ -1,11 +1,11 @@
-﻿<Window x:Class="WpfClient.View.FilmWindow"
+﻿<Window x:Class="WpfClient.View.FilmEditWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:WpfClient.View"
         mc:Ignorable="d"
-        d:DataContext="{d:DesignInstance Type=local:AdminWindow}"
+        d:DataContext="{d:DesignInstance Type=local:FilmEditWindow}"
         
         Title="FilmWindow" Height="450" Width="800">
     <StackPanel Background="AliceBlue">
