+++ WpfClient/View/UserEditWindow.xaml
@@ -1,0 +1,39 @@
+﻿<Window x:Class="WpfClient.View.UserEditWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:WpfClient.View"
+        xmlns:vm="clr-namespace:WpfClient.ViewModel"
+        d:DataContext="{d:DesignInstance Type=local:UserEditWindow}"
+        mc:Ignorable="d"
+        
+        Title="UserWindow" Height="300" Width="500">
+    <Window.DataContext>
+        <vm:ViewModel/>
+    </Window.DataContext>
+    <StackPanel Margin="5" Background="AliceBlue">
+        <TextBlock HorizontalAlignment="Center" FontSize="20">Full info</TextBlock>
+        <UniformGrid>
+            <TextBlock Text="Name: "/>
+            <TextBox x:Name="filmName" Text="{Binding Path=MyUser.Name}"/>
+        </UniformGrid>
+        <UniformGrid>
+            <TextBlock Text="Surname: "/>
+            <TextBox x:Name="FilmYear" Text="{Binding Path=MyUser.Surname}"/>
+        </UniformGrid>
+        <UniformGrid>
+            <TextBlock Text="Email: "/>
+            <TextBox x:Name="FilmDirector" Text="{Binding Path=MyUser.Email}"/>
+        </UniformGrid>
+        <UniformGrid>
+            <TextBlock Text="Phone: "/>
+            <TextBox x:Name="FilmDuration"  Text="{Binding Path=MyUser.Phone}"/>
+        </UniformGrid>
+
+        <UniformGrid VerticalAlignment="Stretch" HorizontalAlignment="Stretch">
+            <Button Width="100" Height="30" Template="{StaticResource  btTemplate}" Click="Button_Click_OK">OK</Button>
+            <Button Width="100" Height="30" Template="{StaticResource  btTemplate}" Click="Button_Click_Cancel">Cancel</Button>
+        </UniformGrid>
+    </StackPanel>
+</Window>
