+++ WpfClient/View/UserWindow.xaml
@@ -4,36 +4,37 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:WpfClient.View"
+        xmlns:sys="clr-namespace:System;assembly=mscorlib"
+
         xmlns:vm="clr-namespace:WpfClient.ViewModel"
-        d:DataContext="{d:DesignInstance Type=local:UserWindow}"
         mc:Ignorable="d"
-        
-        Title="UserWindow" Height="300" Width="500">
+        d:DataContext="{d:DesignInstance Type=vm:ViewModel}"
+        Title="UserWindow" Height="450" Width="800">
     <Window.DataContext>
         <vm:ViewModel/>
     </Window.DataContext>
-    <StackPanel Margin="5" Background="AliceBlue">
-        <TextBlock HorizontalAlignment="Center" FontSize="20">Full info</TextBlock>
-        <UniformGrid>
-            <TextBlock Text="Name: "/>
-            <TextBox x:Name="filmName" Text="{Binding Path=MyUser.Name}"/>
-        </UniformGrid>
-        <UniformGrid>
-            <TextBlock Text="Surname: "/>
-            <TextBox x:Name="FilmYear" Text="{Binding Path=MyUser.Surname}"/>
-        </UniformGrid>
-        <UniformGrid>
-            <TextBlock Text="Email: "/>
-            <TextBox x:Name="FilmDirector" Text="{Binding Path=MyUser.Email}"/>
-        </UniformGrid>
-        <UniformGrid>
-            <TextBlock Text="Phone: "/>
-            <TextBox x:Name="FilmDuration"  Text="{Binding Path=MyUser.Phone}"/>
-        </UniformGrid>
-
-        <UniformGrid VerticalAlignment="Stretch" HorizontalAlignment="Stretch">
-            <Button Width="100" Height="30" Template="{StaticResource  btTemplate}" Click="Button_Click_OK">OK</Button>
-            <Button Width="100" Height="30" Template="{StaticResource  btTemplate}" Click="Button_Click_Cancel">Cancel</Button>
-        </UniformGrid>
-    </StackPanel>
+    <ScrollViewer VerticalScrollBarVisibility="Auto" CanContentScroll="True" Height="Auto">
+        <DockPanel>
+            <Border DockPanel.Dock="Top" Background="Azure">
+                <TextBlock HorizontalAlignment="Center" VerticalAlignment="Center" FontSize="30" Padding="10">
+                    Cinema
+                </TextBlock>
+            </Border>
+            <UniformGrid DockPanel.Dock="Top">
+                <TextBlock>Date: </TextBlock>
+                <DatePicker Name="datePicker1" IsDropDownOpen="False" SelectedDateFormat="Short" DisplayDateStart="{x:Static sys:DateTime.Now}"/>
+            </UniformGrid>
+            <UniformGrid DockPanel.Dock="Top">
+                <TextBlock>Choose cinema hall: </TextBlock>
+                <ComboBox/>
+            </UniformGrid>
+            <UniformGrid DockPanel.Dock="Top">
+                <TextBlock>Choose genre: </TextBlock>
+                <ComboBox/>
+            </UniformGrid>
+            <Border VerticalAlignment="Stretch" Background="LightSkyBlue">
+                <TextBlock>List with films. Use a template for movies with the display of session hours and the ability to choose a session </TextBlock>
+            </Border>
+        </DockPanel>
+    </ScrollViewer>
 </Window>
